Improve error message for header include path conflicts during project generation
Summary:
This used to just throw an IllegalArgumentException that eventually
gets spewn to the command line with huge backtraces (since it's
wrapped in another exception due to lazy loading via a loading cache).

This diff checks for this condition and throws a nicely formatted
HumanReadableException.

Test Plan:
Created 2 headers in different subdirectories, added to an apple_library rule. By default this would generate `libname/foo.h` for both files and result in error.
Run the xcode project generator on this, and see that resulting error message is nice.

Reviewed By: Coneko

fbshipit-source-id: 62272e5