lua: include the native starter in omnibus linking
Summary:
Previously, we excluded native starter deps from omnibus linking.
However, since the native starter can pull in a large number of
deps, it's useful to include it in the link.

This diff accomplishes this by making the `NativeExecutableStarter`
type and instance of `NativeLinkTarget`, so that the omnibus library
can link it as a root.

Test Plan: Updated unittests.

Reviewed By: Coneko

fbshipit-source-id: 2da0584