Update the library for the Eden Thrift API.
Summary:
The new version of the library introduces two changes:
1. The package is renamed from `com.facebook.eden` to `com.facebook.eden.thrift`.
2. The `getSHA1()` API is now a batch API that returns a union type for each path.

This revision updates the `import` statements for the library as well as the
call sites for the `getSHA1()` API. Currently, Buck uses the `getSHA1()` API
only with a single argument. Making full use of the batch API will be done in
a subsequent revision.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 9547a6a