Set intellij python module to use python interpreter named "Python 2"
Summary:
Instead of nothing, or an interpreter named with a specific minor
version and path (which may be different for each machine). Name the
python to use simply the major version.

Users who want the IDE to index python correctly can simply create an
SDK named "Python 2" in their own local intellij config.

Test Plan:
Deleted my local intellij config in ~/Library/Preferences/IdeaC2016.2/... and see that
it doesn't interfere with the rest of the build.

Reviewed By: marcinkosiba

fbshipit-source-id: f07e250