Remove support for Watchman < 3.8.0
Summary:
In his review of D3718760, wez requested that I remove
support for Watchman versions which do not support the extended
`version` command, so we don't need any hard-coded version
support.

With this diff, if the `version` command we send to Watchman
does not include the requested `capabilities`, we assume
Watchman is too old and disable support for it.

Test Plan: Unit tests updated. `ant java-test -Dtest.class=WatchmanTest`

Reviewed By: wez

fbshipit-source-id: 6afdd54