Use new Watchman glob generator when available
Summary:
This builds on a new Watchman feature wez is adding
to greatly improve Watchman glob performance via a new `glob`
generator.

Previously, we used the `path` generator, which created the full list
of files in a directory then filtered them afterwards. For `node.js`
projects with tens of thousands of files in a directory, this could
be very slow.

Now, we use the new `glob` generator when available and pass it
directly the list of glob includes, so it can precisely generate
the list of files needed without iterating over the other files
in the directory.

Depends On D3718760

Test Plan:
Unit tests updated. `buck test //src/com/facebook/buck/json:buck_test`

Ran `buck build buck` with new and old watchman in `PATH`.

Reviewed By: wez

fbshipit-source-id: c205ab3