Place asset catalog contents in Resources
Summary:
For macOS apps, the output of actool should end up in
Foo.app/Contents/Resources, as opposed to in Foo.app/ for iOS apps.
Specifically copy the asset catalog to the resourcesDestinationPath.

Test Plan:
buck build for a macOS app and see its resources under
Contents/Resources/ instead of /

Reviewed By: Coneko

fbshipit-source-id: 55a9db3