Make project command try to pick a better Jdk version for non aggregated modules
Summary:
Now, when we don't aggregate a target, we will try to set the SDK in IntelliJ according to the chosen source-level.
In particular, if all of my tools targets will be Java 8, that will work out of the box correctly.

Test Plan:
Added a unit test

Ran project with targets for a tool that uses Java 8, see that Java 1.8 is defined in the xml file.

Reviewed By: marcinkosiba

fbshipit-source-id: 2aa0781