Make Watchman command timeout configurable
Summary:
The Watchman command timeout is currently hard-coded at 10 seconds
(to match the behavior of Mercurial).

However, on very large repositories, we're seeing the `watchman watch-project`
command (which has to synchronously crawl the entire repository) take longer
than that.

We previously had a config setting `project.watchman_query_timeout_ms` which
we used in the `buck.py` Python parser for Watchman command timeouts; this
extends it to also apply to the initial Watchman command(s) sent to establish
the watch.

This diff:

1. Bumps the default timeout to 45 seconds
2. Makes it configurable

Test Plan:
Unit tests updated.
  ant java-test -Dtest.class=WatchmanTest
  ant java-test -Dtest.class=WatchmanWatcherIntegrationTest

Reviewed By: Coneko

fbshipit-source-id: e890125