cxx: stop using `HasPostBuildSteps` for symlink tree rules
Summary:
The `HasPostBuildSteps` was originally added as a way for rules to
generate some output that wouldn't be cached.  However, there are
some bugs in it's implementation around failure atomicity which can
cause interrupted builds to get stuck in a broken state.

Since we now support selectively disabling build rules, just have
symlink tree rules use that (as this is the original intent).

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: 45c4b24