Support glue dependency in Android native library merging
Summary:
A previous diff added support for merging native libraries in Android
builds.  However, it did not handle the case of a symbol collision,
which is very common with JNI libraries (JNI_OnLoad).  We have a plan to
work around this, which involves adding an extra "glue" dependency to
every library that was the result of merging.

This adds support for that glue.  It is specified as an argument to
android_binary, and is expected to be a cxx_library with
force_static=True.  It is automatically added during linking of merged
libraries, and also added to the "merge_structure" flavor in the target
to ensure that it doesn't leak between apps that use different glue
(though we don't currently plan on doing that).

Test Plan:
New integration tests, including checking that the glue doesn't leak
between apps.

fbshipit-source-id: fb26baa