Support code-generation for Android native library merging
Summary:
In order to support merging of JNI libraries on Android, we need some
representation of the merge structure available at compile time.  The
most convenient format (which avoids the need for an extra native
component) is generated Java code.  Add a new rule that writes out the
merge map in a simple text format, then runs a user-supplied binary rule
to generate the code.  Add a rule to compile the output as Java and
include it in the app.

I think I got the rule keys right, but I'd appreciate a double-check of
that part.

Test Plan: Integration test.

Reviewed By: asp2insp

fbshipit-source-id: 08b0201