Generate files with representative extensions
Summary:
Some parts of Buck's graph transformations rely on the extension of
the source files (e.g. cxx rules). Make sure that the generated
files have the same distribution of extensions as the files in the
original repository.

Test Plan:
Run the script, check that the output passes `buck targets` and the
generated files have correct extensions.

Reviewed By: beefon

fbshipit-source-id: f5569ab