Derive transitive classpaths from transitiveClasspathDeps + immediate classpaths
Summary:
Instead of implementing identical traversal code for finding
transitive dependencies from which to collect classpaths, implement
classpath collection in terms transitive dependencies + classpaths for
each dependency.

This is an intermediate step to removing the
transitiveClasspathEntries multimap entirely.

This diff should be scrutinized with some care, as some of this code
is very old and I have not worked with the Java stuff before.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 511646f