exclude ideabuck module from the main buck project
Summary:
The introduction of ideabuck as a module makes the buck project not build-able unless you also
have the IDEA/IntelliJ SDK. Buttons like `Make project`, `Rebuild project` failed due to the missing
SDK and now you can't easily ran the tests to all the test directory.

Since not a lot of people actually develop this and it also discourages new people from starting to
develop Buck since they have to set up Python, JDKs and the IDEA JDK, which is not trivial, I think
we should remove it from the main project. Maybe we can create a separate project for it.

Test Plan: When you remove the module the project builds without any issues.

Reviewed By: Coneko

fbshipit-source-id: c93b433