store zip locally if endpoint is not accessible
Summary:
This diff makes sure that when buck rage tries to upload to a remote endpoint and fails a backup zip is generated
and stored under buck-out. This way if the user wants, she/he can manually upload it. Also there is a change to
the logic how and in what order we present the response with the server etc. This was so the exception/errors make
sense in the order they arrive.

Test Plan: Edited the test to make sure the zip file is generated if the remote endpoint is not available.

Reviewed By: Coneko

fbshipit-source-id: 3515b1f