haskell: fix build error propagation on haskell binary rule
Summary:
We return a wrapper rule from the haskell binary description which
wraps the link the rule.  However, this rule never added a direct
dependencies on the link rule, which stopped errors from propagating
up and resulted in builds which failed to build but reporting a
successful error code.

Test Plan: added test

fbshipit-source-id: b2a9b4f