fix duplicate entries in version-less libraries for LuaBinary
Summary:
ImmutableSortedMap.Builder throws in case we have multiple identical entries only
when you call .build(). This happened for this a few times for the version-less libraries.
The fix is to use a normal Map and just return a copy of that as an ImmutableSortedMap.

Test Plan: buck test //test/com/facebook/buck/lua/...

Reviewed By: andrewjcg

fbshipit-source-id: 88aa7ca