Fix NPE when an unrecognized event was encountered.
Summary:
Whenever buck got an event (but didn't have an event handler for it), it was
NPE-ing without reporting what event it was, which was cluttering up users'
output.  Instead, let's log that there was an unhandled event (and what its
name was), so that we can later figure out where the event came from.

Test Plan: None -- change to logging that is untested.

fbshipit-source-id: 5e1d8c6