Fix bug in auto of deps in ideabuck.
Summary:
Fixed three-and-a-half distinct bugs in ideabuck related to enabled autodeps:

(1)  When the user has autodeps on, and adds an import where buck
can't identify the source file (most notably for autogenerated
classes or those in third-party jars), ideabuck used to add a
blank target "//:" to the deps.  This has been fixed, in that
if we can't figure out where the file should come from, we don't
add anything.

(2)  When the user imported more than one import statement
by pasting, or when they did a static import, the regex
matching for autodeps wouldn't kick in.  (Fixed.)

(3)  When the user has a dependency in the exported_deps section,
we would duplicate it in deps, even though this is unnecessary,
as all exported_deps are implicitly part of deps.  (Fixed.)

(4)  When the user has autodeps = True enabled in their BUCK file,
the preferred way to add a dependency should be by running
'buck autodeps' to generate a BUCK.autodeps files.  This has
been *somewhat* fixed:  if autodeps is present, the file is
unmodified, but ideabuck neither checks the BUCK.autodeps file
nor does it attempt to run 'buck autodeps'.  (Frankly, it's
unclear to me that a behavior would be appropriate in that
case, as ideabuck's "enable auto deps" isn't synonymous with
'buck autodeps' -- it's just an unfortunate name collision.)

Test Plan: Added unit tests, plus manual testing.

fbshipit-source-id: 7548940