Use smaller dummy data in ArchiveStepIntegrationTest
Summary:
The original allocated 60MiB `(6 characters * 2 bytes per character * 5*2^20)`
right off the bat with the String.repeat, and then when writing it gets encoded
as UTF8, which internally attempts to allocate a 120MiB buffer `(6 characters *
5*2^20 * 4 bytes per character (max utf-8 bytes needed to represent a
characters))`

This was a significant amount of the 256MB usually allocated to integration tests.

Now I reduce the dummy file to 1 MiB, with no intermediate data shuffling by
just writing the bytes without this strings business.

Test Plan: Should not die on CI

Reviewed By: Coneko

fbshipit-source-id: 302e44a