Artificial project: create Android.mk files for ndk_library
Summary:
`ndk_library` requires at least one source to be able to transform
the target into a rule. Let's automatically create an Android.mk
file for each ndk_library. This approach is not entirely correct
since it does not generate any of the other C/C++ sources, but
`ndk_library` is so rarely used that it is not worth it to get it
completely right.

Test Plan:
Run the script, check that `buck targets --show-rulekey` does
not complain about missing source files for ndk_library.

Reviewed By: Coneko

fbshipit-source-id: 0487f43