Switch buck project default to not build targets unless --build is defined
Summary:
Building all modules when doing buck proejct is not a good decision in retrospect. If one wants to see generated code, they are better off building the modules they need and not waiting for all of them in the wrong moment.
The solution is just flipping the default, and requiring someone to add `--build` when they want this. I kept the original flag to now break people's commands in the meantime.

Test Plan: Run `buck project <target>` with and without --build, see that nothing is built without `--build` mentioned.

fbshipit-source-id: a9fe298