new public announcement system
Summary:
This implements the client side of the public announcements system. The goal of this is to deliver messages to users of
Buck. Buck versions and the name of the repository can be specified for better focus and less spam for people that
do not care about the message.

* The change to the thrift API is since these information are not needed.
* Reading the repository from the buckconfig field [cache] repository for now, so we avoid duplication of information.
* Integration test that checks both the behaviour of the thrift API and the passing of the information to the console.
* The message are prefix with `Public Announcements" and have a cyan colouring.

Test Plan:
* An integration test that checks the thrift request and response.
* Manual testing using the appropriate Thrift backend

Reviewed By: ruibm

fbshipit-source-id: c6d7f08