Artificial project: speed up filtering accepted targets
Summary:
Filtering the list of available targets every time is fairly time
consuming. Instead keep a list of already processed targets and
incrementally add to it as more targets are added to the unfiltered
list.

Test Plan:
Run the script, check that it is faster and
`buck targets --show-rulekey` works on the generated project.

Reviewed By: Coneko

fbshipit-source-id: 78d2e98