Add aliases for JDK names
Summary:
Problem: `java_library` with non-default source level is transformed into an iml with JDK that has the same name as source level. For example, level 6 will result in JDK named "6" and 1.6 will result in "1.6". There is often only one JDK for a given level and there is no point in duplicating them.

This change adds support for specifying JDK name aliases in `.buckconfig`. If a source level is found in config, it will be translated, otherwise it will be kept as is (old logic).

This feature can also be used to map all possible source levels to a single JDK.

Test Plan: See integration tests.

Reviewed By: Coneko

fbshipit-source-id: 0f140df